Fit the hpl_glmm_mix_multithread model

Fit the hpl_glmm_mix Stan model and return fit object

Usage

run_hpl_glmm_mix_mt_model(
  method = c("sample", "vb", "pathfinder"),
  run_estimation = 0,
  grainsize = 1,
  G,
  X_g,
  Z_g,
  comps,
  prob,
  y = NULL,
  a_sig2 = NULL,
  b_sig2 = NULL,
  a_sig2_mu = NULL,
  b_sig2_mu = NULL,
  a_mu_offset = NULL,
  b_mu_offset = NULL,
  a_sig2_offset = NULL,
  b_sig2_offset = NULL,
  a_sig2_u = NULL,
  b_sig2_u = NULL,
  normfactors_known = FALSE,
  A_S = NULL,
  B_S = NULL,
  S_DATA = NULL,
  ...
)

Arguments

method

One of `c("sample", "vb", "pathfinder")`

run_estimation

one of `c(0, 1)`; if 0, samples from the prior only and ignores the data

grainsize

Grainsize for multithread

G

Number of groups

X_g

Fixed effects design matrix for a single group

Z_g

Random effects design matrix for a single group

comps

Matrix encoding bernoulli mixture priors

prob

Vector giving the prior probability of each combination of bernoulli components, i.e. of each row in `comps`

y

Numeric vector of output values, sorted by group

a_sig2

vector of shape parameters for Inverse Gamma priors on `sig2`

b_sig2

vector of scale parameters for Inverse Gamma priors on `sig2`

a_sig2_mu

vector of shape parameters for Inverse Gamma priors on `sig2_mu`

b_sig2_mu

vector of scale parameters for Inverse Gamma priors on `sig2_mu`

a_mu_offset

vector of shape parameters for Inverse Gamma priors on `mu_offset`

b_mu_offset

vector of scale parameters for Inverse Gamma priors on `mu_offset`

a_sig2_offset

vector of shape parameters for Inverse Gamma priors on `sig2_offset`

b_sig2_offset

vector of scale parameters for Inverse Gamma priors on `sig2_offset`

a_sig2_u

vector of shape parameters for Inverse Gamma priors on `sig2_u`

b_sig2_u

vector of scale parameters for Inverse Gamma priors on `sig2_u`

normfactors_known

Use fixed normalization factors extrinsic to the model?

A_S

location parameter for `S_PARAM` if `!normfactors_known`

B_S

scale parameter for `S_PARAM` if `!normfactors_known`

S_DATA

Numeric vector of normalization factors for each sample; optional, if `NULL` and `normfactors_known == TRUE`, normalization factors will be estimated by the `TMM` method as described in the `edgeR` package

...

Named arguments to the `sample()` method of CmdStan model objects: <https://mc-stan.org/cmdstanr/reference/model-method-sample.html>

Value

An object of class `CmdStanMCMC`

See also