Set the value of one of the modifiable elements of `standata`

Set the value of one of the modifiable elements of `standata`

Usage

set_standata(
  standata,
  a_sig2 = NULL,
  b_sig2 = NULL,
  a_sig2_mu = NULL,
  b_sig2_mu = NULL,
  a_mu_offset = NULL,
  b_mu_offset = NULL,
  a_sig2_offset = NULL,
  b_sig2_offset = NULL,
  a_sig2_u = NULL,
  b_sig2_u = NULL,
  beta_phi_prior = NULL,
  A_S = NULL,
  B_S = NULL,
  S_DATA = NULL,
  normfactors_known = NULL,
  use_neg_binomial_response = NULL
)

Arguments

standata

The data list to be updated

a_sig2

vector of shape parameters for Inverse Gamma priors on `sig2`

b_sig2

vector of scale parameters for Inverse Gamma priors on `sig2`

a_sig2_mu

vector of shape parameters for Inverse Gamma priors on `sig2_mu`

b_sig2_mu

vector of scale parameters for Inverse Gamma priors on `sig2_mu`

a_mu_offset

vector of shape parameters for Inverse Gamma priors on `mu_offset`

b_mu_offset

vector of scale parameters for Inverse Gamma priors on `mu_offset`

a_sig2_offset

vector of shape parameters for Inverse Gamma priors on `sig2_offset`

b_sig2_offset

vector of scale parameters for Inverse Gamma priors on `sig2_offset`

a_sig2_u

vector of shape parameters for Inverse Gamma priors on `sig2_u`

b_sig2_u

vector of scale parameters for Inverse Gamma priors on `sig2_u`

beta_phi_prior

2-vector giving location and scale parameters for the prior distribution of `beta_phi`

A_S

location parameter for `S_PARAM` if `!normfactors_known`

B_S

scale parameter for `S_PARAM` if `!normfactors_known`

S_DATA

Numeric vector of normalization factors for each sample; optional, if `NULL` and `normfactors_known == TRUE`, normalization factors will be estimated by the `TMM` method as described in the `edgeR` package

normfactors_known

Use fixed normalization factors extrinsic to the model?

use_neg_binomial_response

Use the negative binomial distribution instead of Poisson for the errors?